+ DEV build

+++ src/ladb_opencutlist/cpp/lib/Packy/src/solver.hpp
@@ -718,35 +718,35 @@ namespace Packy {
 
         double to_length_dbl(
                 const Length length,
-                const int8_t precision = 8
+                const int8_t precision = 16
         ) const {
             return round(static_cast<double>(length) / length_truncate_factor_, precision);
         }
 
         static double to_length_dbl(
                 const double length_dbl,
-                const int8_t precision = 8
+                const int8_t precision = 16
         ) {
             return round(length_dbl, precision);
         }
 
         double to_area_dbl(
                 const Area area,
-                const int8_t precision = 8
+                const int8_t precision = 16
         ) const {
             return round(static_cast<double>(area) / (length_truncate_factor_ * length_truncate_factor_), precision);
         }
 
         static double to_area_dbl(
                 const double area_dbl,
-                const int8_t precision = 8
+                const int8_t precision = 16
         ) {
             return round(area_dbl, precision);
         }
 
         static double round(
                 const double value,
-                const int8_t precision = 8
+                const int8_t precision = 16
         ) {
             return std::round(value * std::pow(10, precision)) / std::pow(10, precision);
         }
